Import the Lua 5.3 grammar.
Make the minimal fixes to make it LR compatible, but this is not, as such, a
correct grammar: it does not encode operator precedence correctly, for example.
However, by having this as a base, it should be easier to see what changes are
made to (hopefully) make it fully correct later.