Favour function calls when encountering the grammar's ambiguity.
Lua's grammar is inherently ambiguous. The 5.3 manual defines it thus:

  Function calls and assignments can start with an open parenthesis. This
  possibility leads to an ambiguity in Lua's grammar. Consider the following
  fragment:

     a = b + c
     (print or io.write)('done')

  The grammar could see it in two ways:

     a = b + c(print or io.write)('done')

     a = b + c; (print or io.write)('done')

  The current parser always sees such constructions in the first way,
  interpreting the open parenthesis as the start of the arguments to a call.

Rewriting the grammar to favour the first way is extremely difficult (too
difficult for me). But we can make use of the fact that the current grammar
contains 1 shift/reduce and 1 reduce/reduce conflict, both of which relate to
the above. The shift/reduce is automatically resolved (using Yacc's normal
rules) in the way we want. However, the reduce/reduce conflict is between stat
and prefixexp, favouring stat, which means that we parse the above example in
the second way.

Fortunately we can fix this easily by simply moving the prefixexp rule above the
stat rule in the .y file: this resolves the reduce/reduce conflict in favour of
prefixexp, meaning that we parse the above example in the first way.